refactor: extract side effects to src/lib/render.ts
